requirements.txt
click==7.1.2
Flask==1.1.1
Flask-Cache==0.13.1
Flask-Caching==1.9.0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.1
Pillow==7.2.0
Werkzeug==1.0.1
